pyproject.toml
[tool.poetry]
name = "2lab"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.105.0"
uvicorn = "^0.24.0.post1"
alembic = "^1.13.1"
psycopg = "^3.1.16"
asyncpg = "^0.29.0"
postgres = "^4.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
